(26/05/25)
Made some small changes to the shape and size of the buttons and labels of the program.

I also did some testing on the functionality of using japanese words and inputs into godot and it worked well, but I have since reverted the changes as they were temporary for testing purposes.