5/06/25
I made some changes to the flashcard system with the positioning and such of the stuff, however it didn't really work as I expected so I will have to revisit that and change up some of the code and node types for the flashcard system. I also created a new scene for the space invaders game portion of the game so I will have that set up for when I need it.