10/06/25
Added some minor code to the bullet and shooting code to make it so the bullets shot by the player remove themselves from the games tree and memory 2 seconds after being shot. This means that the game doesn't get slowed down by a lot of memory being taken up by the bullets shot. I also assigned some layers to the game to help with collision down the line. I also set up a scene and class for enemies and their spawning to work on later.