26/06/25
I renamed a lot of the assets for the game to more suitable names and I also added enemy shooting to the game with a framework to change their stats. My enemy bullets currently do not do anything to the player. I probably wont finish all of my sprint goals this sprint but I am making good progress towards all of them.