22/07/2025
I added a health and basic game over system to the player when it gets hit 3 times, I also have it set up in a way so that the health of the player can be modified easily. I added a function for when all of the enemies are destroyed where the game stops, a very basic victory condition screen effectively. I also added a basic function for testing at the moment where if the player losses all of their health the game transitions to the flashcard portion of the game, it is very basic in its current state but it works as I have intended it to.