29/07/2025
I played around with some of the logic for detection of enemies being destroyed and it worked, so I was able to link that to the powerup logic and I got my powerups working. The current powerups increase player speed, increase fire rate, and increase player health, and they are all working as expected.