01/08/2025
I added waves to the game so that more enemies spawn once the current ones have been killed. The switch between waves is currently a bit buggy but they are functional so I will clean them up a bit. I also added a wave counter next to the health counter which works as expected, showing the current wave number and increase when moving onto the next wave of enemies.