04/08/2025
I added waves into the game so now they work as intended with powerups, enemies spawning, and everything else related to it. I also added difficulty scaling to the game so that as the game goes on it progressively gets harder. The game gets harder by spawning more enemies and having the enemies move faster. I should be able to focus on linking the flashcard system to the game for the rest of the week as I got everything else I wanted to do this sprint working.