05/08/2025
I polished up the difficulty scaling between waves so that the game doesn't become impossible but gets noticeably harder as it was too easy before. I started to link the flashcards to the game by making a main script on the main node to change when different parts of the game should be toggled and active.