08/08/2025
I linked the flashcard system to the game fully meaning that after a wave of enemies, the UI elements for the game are hidden and de-activated and the UI elements for the flashcards are shown and they can be interacted with. Once the player has correctly answered 3 flashcards, they are returned to the game and the next wave begins.