28/08/2025
I added instructions to the start page of the game, I also added labels on all of the buttons so that users can tell what they do. I increased the font size of many labels to make them easier to read. I made a small change to the flashcard system where the input box for answers disappears and is replaced with the correct or incorrect label depending on if the answer given was correct or not.

I got rid of any temporary functions that printed values to terminal as they were used for testing but are now not needed.